requirements.txt
onnxruntime>=1.16.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
paddlepaddle>=2.6.2
paddleocr>=2.7.0
PyYAML>=6.0
streamlit>=1.28.0
